(refactor) 3단계 리펙토링 - dto로 변경
